requirements.txt
Django==5.0.1
Pillow==10.1.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-allauth==0.58.2
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0